+++ src/callbacks/postNewLevel.ts
@@ -4,6 +4,7 @@ import * as streakText from "../features/mandatory/streakText";
 import * as tempMoreOptions from "../features/mandatory/tempMoreOptions";
 import * as openHushDoor from "../features/optional/bosses/openHushDoor";
 import { extraStartingItemsPostNewLevel } from "../features/optional/gameplay/extraStartingItems/callbacks/postNewLevel";
+import * as fastVanishingTwin from "../features/optional/quality/fastVanishingTwin";
 import { showDreamCatcherItemPostNewLevel } from "../features/optional/quality/showDreamCatcherItem/callbacks/postNewLevel";
 import * as silenceMomDad from "../features/optional/sound/silenceMomDad";
 import { racePostNewLevel } from "../features/race/callbacks/postNewLevel";
@@ -31,7 +32,8 @@ export function main(): void {
 
   // QoL
   openHushDoor.postNewLevel();
-  showDreamCatcherItemPostNewLevel();
+  showDreamCatcherItemPostNewLevel(); // 566
+  fastVanishingTwin.postNewLevel(); // 697
 
   // Sounds
   silenceMomDad.postNewLevel();